+++ src/panels/energy/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_ENERGY_COLLECTION_KEY = "energy_dashboard";
+export const DEFAULT_POWER_COLLECTION_KEY = "energy_dashboard_now";
